requirements.txt
asgiref==3.7.2
boto==2.49.0
boto3==1.34.0
botocore==1.34.0
certifi==2023.11.17
charset-normalizer==3.3.2
crispy-bootstrap5==2023.10
Django==5.0
django-cors-headers==4.3.1
django-crispy-forms==2.1
django-filter==23.5
django-storages==1.14.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
idna==3.6
jmespath==1.0.1
Markdown==3.5.1
Pillow==10.1.0
psycopg2==2.9.9
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
s3transfer==0.9.0
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.7
whitenoise==6.6.0
